LayoutTests/imported/w3c/web-platform-tests/html/user-activation/message-event-init.tentative-expected.txt


FAIL MessageEventInit user activation set assert_equals: userActivation attribute expected (object) object "[object UserActivation]" but got (undefined) undefined
FAIL MessageEventInit user activation not set assert_equals: userActivation attribute expected (object) null but got (undefined) undefined
